requirements/production.txt
Django==1.9.7
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
PyJWT==1.4.0
